Before merging there is a bug that means that whatever is first set is there forever
