+++ SubtitleWindow.py
@@ -2,10 +2,8 @@ import sys
 from PyQt5.QtWidgets import QApplication, QLabel, QMenu, QAction, QMainWindow
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QFontMetrics
-from pathlib import Path
-import Profiles
-import yaml
-import os
+from Profiles import Profiles
+from SettingsWindow import SettingsWindow
 
 class SubtitleWindow(QMainWindow):
     def __init__(self):
@@ -13,7 +11,7 @@ class SubtitleWindow(QMainWindow):
         #recieves userProfiles if unsuccessful then quits the application
         self.profiles = Profiles.getUserProfiles()
         if self.profiles['Current'] is None:
-            self.userSelector = Profiles.UserSelector(self.profiles)
+            self.userSelector = SettingsWindow(self.profiles)
             self.userSelector.exec_()
         self.profiles = Profiles.getUserProfiles()
         if self.profiles['Current'] is None:
@@ -34,6 +32,7 @@ class SubtitleWindow(QMainWindow):
         self.translucentMode = True
         # set label which contains the subtitles
         self.label = QLabel(self)
+        self.label.setText("[Subtitles]")
         self.label.setStyleSheet(self.css)
         self.label.setGeometry(0, 0, self.width(), self.height())
 
@@ -76,9 +75,13 @@ class SubtitleWindow(QMainWindow):
         else:
             translucent.setText("Disable Translucent Background")
         contextMenu.addAction(translucent)
+        settingsAction = QAction("Settings",self)
+        settingsAction.triggered.connect(self.openSettingsMenu)
+        contextMenu.addAction(settingsAction)
         #quit
         quit = QAction("Quit")
         contextMenu.addAction(quit)
+        
         action = contextMenu.exec_(self.mapToGlobal(event.pos()))
         if action == quit:
             self.close()
@@ -88,7 +91,25 @@ class SubtitleWindow(QMainWindow):
         self.translucentMode = not self.translucentMode
         self.setAttribute(Qt.WA_TranslucentBackground,self.translucentMode)
         self.update()
-        
+    
+    def openSettingsMenu(self):
+        self.hide()
+        self.profiles = Profiles.getUserProfiles()
+        self.userSelector = SettingsWindow(self.profiles)
+        self.userSelector.exec_()
+        if self.profiles['Current'] is None:
+            self.userSelector = SettingsWindow(self.profiles)
+            self.userSelector.exec_()
+        self.profiles = Profiles.getUserProfiles()
+        if self.profiles['Current'] is None:
+            sys.exit()
+        else:
+            with open(self.profiles['Users'][self.profiles['Current']],"r") as userFile: #Opens Current User File I know its cursed plz somebody fix
+                self.css = userFile.read()
+        self.label.setText("[Subtitles]")
+        self.label.setStyleSheet(self.css)
+        self.show()
+
     # if window is resized ensures that the label will stay the same size
     def resizeEvent(self, event):
         self.label.setGeometry(0,0,self.width(),self.height())
